requirements.txt
aiohttp[speedups]=3.7.4.post0
Jinja2==3.0.1
python-resize-image==1.1.19